+++ code/core/src/main/java/com/google/cloud/broker/database/backends/AbstractDatabaseBackend.java
@@ -15,7 +15,7 @@ import com.google.cloud.broker.settings.AppSettings;
 import com.google.cloud.broker.database.DatabaseObjectNotFound;
 import com.google.cloud.broker.database.models.Model;
 import com.google.cloud.broker.utils.InstanceUtils;
-
+import com.google.cloud.broker.checks.CheckResult;
 
 public abstract class AbstractDatabaseBackend {
 
@@ -25,6 +25,7 @@ public abstract class AbstractDatabaseBackend {
     public abstract void save(Model model);
     public abstract void delete(Model model);
     public abstract void initializeDatabase();
+    public abstract CheckResult checkConnection();
 
     public static AbstractDatabaseBackend getInstance() {
         String className = AppSettings.getInstance().getString(AppSettings.DATABASE_BACKEND);