+++ code/extensions/caching/cloud-datastore/src/main/java/com/google/cloud/broker/caching/remote/CloudDatastoreCache.java
@@ -11,13 +11,17 @@
 
 package com.google.cloud.broker.caching.remote;
 
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.Lock;
 
+import com.google.cloud.datastore.*;
+
 import com.google.cloud.broker.settings.AppSettings;
 import com.google.cloud.broker.utils.TimeUtils;
-import com.google.cloud.datastore.*;
+import com.google.cloud.broker.checks.CheckResult;
 
 public class CloudDatastoreCache extends AbstractRemoteCache {
 
@@ -191,6 +195,22 @@ public class CloudDatastoreCache extends AbstractRemoteCache {
         DatastoreLock lock = new DatastoreLock(lockName);
         lock.lock();
         return lock;
+    }
+
+    @Override
+    public CheckResult checkConnection() {
+        try {
+            Datastore datastore = getService();
+            Query<Entity> query = Query.newEntityQueryBuilder()
+                .setKind("ABCDEFGHIJ1234567890")  // Arbitrary fictitious Kind
+                .build();
+            datastore.run(query);
+            return new CheckResult(true);
+        } catch(Exception e) {
+            StringWriter sw = new StringWriter();
+            e.printStackTrace(new PrintWriter(sw));
+            return new CheckResult(false, sw.toString());
+        }
     }
 
 }