+++ CHANGES.md
@@ -4,6 +4,8 @@
 - Added new base client library: `broker-client-lib`.
 - Renamed `broker-connector` package to `broker-hadoop-connector`.
 - Added `PingServer` command.
+- Added system check when the broker server starts to test connectivity
+  with database, cache, and encryption backends.
 
 ## 0.9.2 (March 9, 2020)